Clear object cache for orders + products #5041

+++ includes/wc-order-functions.php
@@ -340,12 +340,26 @@ function wc_delete_shop_order_transients( $post_id = 0 ) {
 		'woocommerce_processing_order_count'
 	);
 
+	// Clear transients for which we don't have the name
+	if ( wp_using_ext_object_cache() ) {
+		global $wp_object_cache;
+
+		if ( isset( $wp_object_cache['transient'] ) ) {
+			$keys = array_keys( $wp_object_cache['transient'] );
+			foreach ( $keys as $key ) {
+				if ( 'wc_report_' === substr( $key, 0, 10 ) ) {
+					$transients_to_clear[] = $key;
+				}
+			}
+		}
+	} else {
+		$wpdb->query( "DELETE FROM `$wpdb->options` WHERE `option_name` LIKE ('_transient_wc_report_%') OR `option_name` LIKE ('_transient_timeout_wc_report_%')" );
+	}
+
+	// Clear transients where we have names
 	foreach( $transients_to_clear as $transient ) {
 		delete_transient( $transient );
 	}
-
-	// Clear transients for which we don't have the name
-	$wpdb->query( "DELETE FROM `$wpdb->options` WHERE `option_name` LIKE ('_transient_wc_report_%') OR `option_name` LIKE ('_transient_timeout_wc_report_%')" );
 
 	do_action( 'woocommerce_delete_shop_order_transients', $post_id );
 }
